move functions to Utils
